schematelemetry: add unit tests for descriptor redaction
This commit moves code around to make it testable but otherwise does not
alter any functionality.

Informs #85787.

Release note: none
