+++ pkg/sql/catalog/schematelemetry/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "schematelemetry",
     srcs = [
+        "redact.go",
         "scheduled_job_executor.go",
         "schema_telemetry_event.go",
         "schema_telemetry_job.go",
@@ -48,18 +49,23 @@ go_test(
     name = "schematelemetry_test",
     srcs = [
         "main_test.go",
+        "redact_test.go",
         "schema_telemetry_test.go",
     ],
     args = ["-test.timeout=295s"],
     deps = [
+        ":schematelemetry",
         "//pkg/base",
         "//pkg/jobs",
         "//pkg/jobs/jobstest",
+        "//pkg/keys",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/sql",
+        "//pkg/sql/catalog/desctestutils",
         "//pkg/sql/catalog/schematelemetry/schematelemetrycontroller",
+        "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/sem/builtins/builtinconstants",
         "//pkg/sql/sem/tree",
         "//pkg/testutils/serverutils",
